+++ src/panel_material_ui/pane/base.py
@@ -6,6 +6,7 @@ from typing import Any
 import param
 from panel.links import Callback
 from panel.pane import HoloViews, Markdown
+from panel.viewable import Child
 
 from ..base import COLORS, MaterialComponent, ThemedTransform, TooltipTransform
 from ..widgets import DatetimeInput, DiscreteSlider, EditableFloatSlider, EditableIntSlider, FloatSlider, IntSlider, Select
@@ -309,6 +310,71 @@ class Skeleton(MaterialPaneBase):
     )
 
     _esm_base = "Skeleton.jsx"
+
+
+class Tooltip(MaterialComponent):
+    """
+    The `Tooltip` displays informative text when users hover over, focus
+    on, or tap a child element. It wraps a single child component and
+    shows a configurable tooltip label.
+
+    :References:
+
+    - https://panel-material-ui.holoviz.org/reference/panes/Tooltip.html
+    - https://mui.com/material-ui/react-tooltip/
+
+    :Example:
+
+    >>> Tooltip(Button(label="Delete"), title="Remove this item")
+    """
+
+    arrow = param.Boolean(default=False, doc="""
+        Whether the tooltip has an arrow indicating the element it
+        refers to.""")
+
+    describe_child = param.Boolean(default=False, doc="""
+        Whether the tooltip acts as an accessible description rather
+        than a label. Use when the child already has a visible label
+        and the tooltip provides supplementary information.""")
+
+    enter_delay = param.Integer(default=100, bounds=(0, None), doc="""
+        The number of milliseconds to wait before showing the tooltip.
+        This can help avoid tooltips appearing on quick mouse passes.""")
+
+    follow_cursor = param.Boolean(default=False, doc="""
+        Whether the tooltip follows the cursor position.""")
+
+    leave_delay = param.Integer(default=0, bounds=(0, None), doc="""
+        The number of milliseconds to wait before hiding the tooltip.""")
+
+    object = Child(doc="""
+        The child component to wrap with the tooltip.""")
+
+    open = param.Boolean(default=None, allow_None=True, doc="""
+        Explicitly control whether the tooltip is open. When None,
+        the tooltip is managed automatically on hover/focus. Set to
+        True or False for programmatic control.""")
+
+    placement = param.Selector(
+        default="bottom",
+        objects=[
+            "bottom-end", "bottom-start", "bottom",
+            "left-end", "left-start", "left",
+            "right-end", "right-start", "right",
+            "top-end", "top-start", "top",
+        ],
+        doc="""
+        The placement of the tooltip relative to the child element.""")
+
+    title = param.String(default="", doc="""
+        The text to display inside the tooltip.""")
+
+    _esm_base = "Tooltip.jsx"
+
+    def __init__(self, object=None, **params):
+        if object is not None:
+            params["object"] = object
+        super().__init__(**params)
 
 
 class Typography(MaterialPaneBase, Markdown):
